refactor: pull out home page
