+++ src/components/Panel3/index.js
@@ -1,0 +1,1 @@
+export { default } from './Panel3';
